Rename lib/tasks/sass.js to lib/tasks/test-sass-compilation.js
