+++ lib/tasks/test.js
@@ -4,7 +4,7 @@ const commandLine = require('../helpers/command-line');
 const files = require('../helpers/files');
 const pa11y = require('./pa11y');
 const karma = require('./karma');
-const sass = require('./sass');
+const getSassCompilationTests = require('./test-sass-compilation');
 const testSass = require('./test-sass');
 
 const npmTest = function (config) {
@@ -18,11 +18,10 @@ module.exports = function (cfg) {
 
 	const Listr = require('listr');
 
-	const compilationTests = sass(config);
-	return compilationTests.then((tests) => {
+	return getSassCompilationTests(config).then((compilationTests) => {
 		return new Listr(
 			[
-				...tests,
+				...compilationTests,
 				testSass(config),
 				{
 					title: 'Executing `npm test`',
